scale to 24dp

+++ library/src/main/java/com/balysv/materialmenu/MaterialMenuView.java
@@ -64,7 +64,7 @@ public class MaterialMenuView extends View implements MaterialMenu {
 
         try {
             int color = attr.getColor(R.styleable.MaterialMenuView_mm_color, DEFAULT_COLOR);
-            int scale = attr.getInteger(R.styleable.MaterialMenuView_mm_scale, DEFAULT_SCALE);
+            float scale = attr.getFloat(R.styleable.MaterialMenuView_mm_scale, DEFAULT_SCALE);
             int transformDuration = attr.getInteger(R.styleable.MaterialMenuView_mm_transformDuration, DEFAULT_TRANSFORM_DURATION);
             int pressedDuration = attr.getInteger(R.styleable.MaterialMenuView_mm_pressedDuration, DEFAULT_PRESSED_DURATION);
             Stroke stroke = Stroke.valueOf(attr.getInteger(R.styleable.MaterialMenuView_mm_strokeWidth, 0));
